Refactor in preparation for shifting work to arq
Extract the background processing from the process context, since
we will need to have a process context for workers that doesn't
also do background processing. Separate results processing into its
own service, used by both the query service and the query monitor,
so that result processing workers will be able to use it directly.
