+++ src/qservkafka/storage/state.py
@@ -33,27 +33,6 @@ class QueryStateStore:
         self._storage = storage
         self._logger = logger
 
-    async def add_query(
-        self, query_id: int, job: JobRun, status: AsyncQueryStatus
-    ) -> None:
-        """Add a record for a newly-created query.
-
-        Parameters
-        ----------
-        query_id
-            Qserv query ID.
-        job
-            Original job request.
-        status
-            Initial query status.
-        """
-        query = await self._storage.get(str(query_id))
-        if query:
-            msg = "Duplicate query ID, replacing old query record"
-            self._logger.error(msg, query_id=query_id)
-        query = Query(query_id=query_id, job=job, status=status)
-        await self._storage.store(str(query_id), query)
-
     async def delete_query(self, query_id: int) -> None:
         """Delete a query from storage.
 
@@ -93,10 +72,10 @@ class QueryStateStore:
         """
         return await self._storage.get(str(query_id))
 
-    async def update_status(
+    async def store_query(
         self, query_id: int, job: JobRun, status: AsyncQueryStatus
     ) -> None:
-        """Add a record for a newly-created query.
+        """Add or update a record for an in-progress query.
 
         Parameters
         ----------
@@ -111,6 +90,5 @@ class QueryStateStore:
         if query:
             query.status = status
         else:
-            # This case shouldn't happen but we may as well handle it.
             query = Query(query_id=query_id, job=job, status=status)
         await self._storage.store(str(query_id), query)
